requirements.txt
duckdb==1.1.3
gitdb==4.0.11
GitPython==3.1.41
numpy==2.2.0
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
setuptools==69.0.3
six==1.17.0
smmap==5.0.1
tzdata==2024.2
